Update README.md
Fixed some formatting issues.

In particular javascript code blocks in bullet lists are now rendering correctly.  Git hub markdown is apparently inconsistent with how html and javascript code blocks are render and what was working for the html blocks was failing for the javascript blocks.

I have also indented much of the text in the bullet points. For similar reasons.