Make templater.clear a bit faster
By using a range, we can improve performance of templater.clear() by
roughly 25% (tested in Chrome). I decided to make this perf enhancement
opt-in to browsers supporting the `document.createRange` method. It's
not supported in IE<9, but most modern browsers should have it.

On a list with 1000 items, this change took down the time it takes for
templater.clear() from roughly 40ms to <30ms. In total, the whole
search execution takes about 160ms, so this shaves off ~6% of execution
time.