Use document fragment to speed up list.update()
By adding DOM nodes one by one, we're causing multiple reflows in the
browser. This can have bad effects on performane when lists are longs.
By adding DOM nodes to a fragment we can limit the number of reflows to
one. This cuts execution time for list.update() in half (40ms to 20ms)
on a list with 1000 items.

While I was in the code, I noticed a misspelled comment that I fixed.