Added pagination option to include next/prev buttons.
Removed the e.preventDefault, and changed pagination to use html button tags instead of anchor tags with "#" as the href. As these are not links, a button is a semantically sound approach to handling the pagination and avoids issues of scrolling to the top of the page whenever a page button is clicked.